Create RemoteUriResolver, untested
Would like to test it on a repo with a file, it should check the config and then download it. Tests for this are probably warranted? Should this move to the base library?
After this is confirmed working, we need to actually perform the Android.mk edits and c_cpp_properties.json and also get the .so